requirements.txt
Django==3.2.4
pdfminer.six
crispy-bootstrap5